rename header types
